requirements.txt
azureml-dataprep
gensim
numpy
keras
tensorflow==2.1.0
tensorflow-hub
transformers
sklearn
